Add files via upload
Code Challenge:
In line 13, between the div tags, add a JSX span element (with both start and end tags). Embed the property "bootcamp" stored in the state of this component inside the span element's start and end tags.
Add another JSX span element, and embed the variable "course" inside this one.
In the CSS section, there are two classes: .blue-box and .orange-box. Apply the .blue-box class to the first JSX span element, and the .orange-box class to the second.
NOTE: In this and future code challenges that use Codepen, you may notice that class components use the syntax "extends React.Component" whereas in your nucampsite project, you will use the syntax "extends Component".